Add stdc++fs as a target link library for clang compiler on linux (#144)
* Add stdc++fs as a target link library for clang compiler on linux

This change is needed to compile pluginlib package and packages which
depend on it on linux clang compiler

* Add and export stdc++fs lib for linux platform

* Add c++fs library link on all non-Apple *nix platforms.

This change is necessary for building with Clang on Linux. The use of
`UNIX AND NOT APPLE` is cribbed from a bunch of StackOverflow and CMake
mailing list examples to cover non-Apple *nix cases which for ROS 2
today means Linux, basically. But in the future it could also mean
NetBSD or FreeBSD, QNX, or NuttX as long as they provide libc++fs.